+++ src/app/browse-products.tsx
@@ -1,36 +1,84 @@
+'use client';
+
 import styles from '@/styles/app/page.module.scss';
 import { BrowseProduct } from '@/app/browse-product';
 import { OtherProductType, ProductType } from '@/types/main-page-types';
 import { BreakLine } from '@/components/common/break-line';
-import { Product } from '@/app/product';
+import { useState } from 'react';
+import { ProductPagination } from '@/app/product-pagination';
+import { OtherProducts } from '@/app/other-products';
 
 export function BrowseProducts() {
+    const [page, setPage] = useState<number>(1);
+    const productsPerPage = 3;
+
+    const filterProducts = (otherProducts: OtherProductType[]): OtherProductType[] => {
+        return otherProducts.filter((product, index) => {
+            const startIndex = (page - 1) * productsPerPage;
+            const endIndex = startIndex + productsPerPage;
+
+            return index >= startIndex && index < endIndex;
+        });
+    };
+
     const products: ProductType[] = [
-        { name: 'Sport', imageUrl: '' },
-        { name: 'Life', imageUrl: '' },
-        { name: 'Health', imageUrl: '' }
+        { name: 'Sport', imageUrl: '/pages/main/sport-category.jpg' },
+        { name: 'Life', imageUrl: '/pages/main/life-category.jpg' },
+        { name: 'Health', imageUrl: '/pages/main/health-category.jpg' }
     ];
 
     const otherProducts: OtherProductType[] = [
         {
             name: 'Masaż fizjoterapeutyczny',
-            imageUrl: '',
+            imageUrl: '/pages/main/massage.jpg',
             location: 'Kraków, w. Małopolskie',
             address: 'al. Pokoju 16',
             priceFrom: 49,
             currency: 'zł'
         },
         {
-            name: 'Masaż fizjoterapeutyczny',
-            imageUrl: '',
+            name: 'Masaż fizjoterapeutyczny 2',
+            imageUrl: '/pages/main/massage.jpg',
             location: 'Kraków, w. Małopolskie',
             address: 'al. Pokoju 16',
             priceFrom: 49,
             currency: 'zł'
         },
         {
-            name: 'Masaż fizjoterapeutyczny',
-            imageUrl: '',
+            name: 'Masaż fizjoterapeutyczny 3',
+            imageUrl: '/pages/main/massage.jpg',
+            location: 'Kraków, w. Małopolskie',
+            address: 'al. Pokoju 16',
+            priceFrom: 49,
+            currency: 'zł'
+        },
+        {
+            name: 'Masaż fizjoterapeutyczny 4',
+            imageUrl: '/pages/main/massage.jpg',
+            location: 'Kraków, w. Małopolskie',
+            address: 'al. Pokoju 16',
+            priceFrom: 49,
+            currency: 'zł'
+        },
+        {
+            name: 'Masaż fizjoterapeutyczny 5',
+            imageUrl: '/pages/main/massage.jpg',
+            location: 'Kraków, w. Małopolskie',
+            address: 'al. Pokoju 16',
+            priceFrom: 49,
+            currency: 'zł'
+        },
+        {
+            name: 'Masaż fizjoterapeutyczny 6',
+            imageUrl: '/pages/main/massage.jpg',
+            location: 'Kraków, w. Małopolskie',
+            address: 'al. Pokoju 16',
+            priceFrom: 49,
+            currency: 'zł'
+        },
+        {
+            name: 'Masaż fizjoterapeutyczny 7',
+            imageUrl: '/pages/main/massage.jpg',
             location: 'Kraków, w. Małopolskie',
             address: 'al. Pokoju 16',
             priceFrom: 49,
@@ -48,11 +96,13 @@ export function BrowseProducts() {
             </div>
             <BreakLine />
             <div className={styles['other-products']}>
-                <div className={styles['other-products__container']}>
-                    {otherProducts.map(product => <Product product={product} key={product.name} />)}
-                </div>
-                {/* // TODO Pagination */}
-                <div className={styles['other-products__pagination-container']}>X pagination X</div>
+                <OtherProducts products={filterProducts(otherProducts)} />
+                <ProductPagination
+                    setPage={setPage}
+                    page={page}
+                    productCount={otherProducts.length}
+                    productsPerPage={productsPerPage}
+                />
             </div>
             <></>
         </div>