+++ src/app/product.tsx
@@ -9,7 +9,9 @@ export function Product({ product }: ProductProps) {
     return (
         <div className={styles['product']}>
             <div className={styles['product__image-container']}>
-                <div className={styles['product__image']}></div>
+                <div className={styles['product__image']}>
+                    <img src={product.imageUrl} alt={product.name} />
+                </div>
             </div>
             <div className={styles['product__info-container']}>
                 <div>
@@ -18,7 +20,7 @@ export function Product({ product }: ProductProps) {
                     <div className={styles['product__info-address']}>{product.address}</div>
                 </div>
                 <div>
-                    <button>od {product.priceFrom} {product.currency}</button>
+                    <span>od {product.priceFrom} {product.currency}</span>
                     <p className={styles['product__info-summary']}>Za całość zapłać punktami</p>
                 </div>
             </div>