+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-04-08
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `Button` common components with sizes: `small`, `normal` and `big`.
+- [Marek Kurańda](https://github.com/mjkuranda): Missing images to main page.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): styling of some components attached to main page.
+
 ## [0.1.0] - 2024-04-08
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): initial version of main page.